# docs/swagger.yaml
basePath: /v1
definitions:
  api.createIncomeRequest:
    properties:
      balance:
        description: |-
          Amount of the income.
          Required: true
          example: 1000
          in: body
        type: number
      payee:
        description: |-
          Payee of the income.
          Required: true
          example: john_doe
          in: body
        type: string
      project_id:
        description: |-
          ProjectID of the income.
          Required: true
          swagger:strfmt uuid
          example: 123e4567-e89b-12d3-a456-426614174000
          in: body
        type: string
    required:
    - balance
    - payee
    - project_id
    type: object
  api.createLoanRequest:
    properties:
      amount:
        description: |-
          Amount of the loan.
          Required: true
          example: 1000
          in: body
        type: number
      borrower:
        description: |-
          Borrower of the loan.
          Required: true
          example: john_doe
          in: body
        type: string
      subject:
        description: |-
          Subject of the loan.
          Required: true
          example: loan1
          in: body
        type: string
    required:
    - amount
    - borrower
    - subject
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
  api.listRequest:
    properties:
      page_id:
        description: |-
          PageID is the page number.
          Required: true
          example: 1
          in: body
          minimum: 1
        minimum: 1
        type: integer
      page_size:
        description: |-
          PageSize is the number of projects per page.
          Required: true
          example: 5
          in: body
          minimum: 5
          maximum: 10
        maximum: 100
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
...
  api.searchRequest:
    properties:
      page_id:
        description: |-
          PageID is the page number.
          Required: true
          example: 1
          in: body
          minimum: 1
        minimum: 1
        type: integer
      page_size:
        description: |-
          PageSize is the number of projects per page.
          Required: true
          example: 5
          in: body
          minimum: 5
          maximum: 10
        maximum: 10
        minimum: 5
        type: integer
      query:
        description: |-
          Query is the search query.
          Required: true
          example: project1
          in: body
        type: string
    required:
    - page_id
    - page_size
    - query
    type: object
...
  db.Income:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      payee:
        type: string
      project_id:
        type: string
      updated_at:
        type: string
    type: object
  db.Loan:
    properties:
      amount:
        type: number
      borrower:
        type: string
      created_at:
        type: string
      id:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: lushenle@gmail.com
    name: API Support
    url: http://localhost:8080/api/v1/support
  description: This is a sample server for PLAM API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://localhost:8080/api/v1/terms/
  title: PLAM API
  version: "1.0"
paths:
  /incomes:
    post:
      consumes:
      - application/json
      description: Create a new income.
      parameters:
      - description: Create Income Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createIncomeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Income created
          schema:
            $ref: '#/definitions/db.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create an income
      tags:
      - incomes
  /incomes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an income by ID.
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted income
          schema:
            $ref: '#/definitions/db.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an income
      tags:
      - incomes
    get:
      consumes:
      - application/json
      description: Get an income by ID.
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Income found
          schema:
            $ref: '#/definitions/db.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get an income
      tags:
      - incomes
  /incomes/all:
    post:
      consumes:
      - application/json
      description: List all incomes.
      parameters:
      - description: List Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.listRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Income'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all incomes
      tags:
      - incomes
  /incomes/search:
    post:
      consumes:
      - application/json
      description: Search incomes by payee.
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.searchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Incomes found
          schema:
            items:
              $ref: '#/definitions/db.Income'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search incomes
      tags:
      - incomes
  /loans:
    post:
      consumes:
      - application/json
      description: Create a new loan.
      parameters:
      - description: Create Loan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createLoanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Loan created
          schema:
            $ref: '#/definitions/db.Loan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a loan
      tags:
      - loans
  /loans/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a loan by ID.
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lone deleted
          schema:
            $ref: '#/definitions/db.Loan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a loan
      tags:
      - loans
    get:
      consumes:
      - application/json
      description: Get a loan by ID.
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Loan found
          schema:
            $ref: '#/definitions/db.Loan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a loan
      tags:
      - loans
  /loans/all:
    post:
      consumes:
      - application/json
      description: List all loans.
      parameters:
      - description: List Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.listRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of loans
          schema:
            items:
              $ref: '#/definitions/db.Loan'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all loans
      tags:
      - loans
  /loans/search:
    post:
      consumes:
      - application/json
      description: Search loans by borrower.
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.searchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of loans
          schema:
            items:
              $ref: '#/definitions/db.Loan'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search loans
      tags:
      - loans
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"